+++ crates/ruma-events/src/room.rs
@@ -21,6 +21,8 @@ pub mod encryption;
 pub mod guest_access;
 pub mod history_visibility;
 pub mod join_rules;
+#[cfg(feature = "unstable-msc4334")]
+pub mod language;
 pub mod member;
 pub mod message;
 pub mod name;